+++ R/shortcut.R
@@ -7,10 +7,19 @@
 #' on. The shortcut forms skip the construction of the
 #' intermediate [quotation] objects.
 #'
-#' As throughout this package, the functions with bare names will
-#' quote their first argument literally. Their counterparts, with
-#' underscores `_` at the end, take a character vector or list of
-#' names.  and an environment to look in.
+#' Throughout this package, some functions come in two forms, a "bare"
+#' version which quote their first argument literally, and a
+#' normally-evaluating version with a trailing underscore in its
+#' name. So `is_forced(x)` chiecks whether "x" is a missing variable,
+#' while `is_forced_(x, environment())` checks whether "x"
+#' contains the _name_ of another variable which is missing. The
+#' following are all equivalent:
+#'
+#' * `arg_env(x)`
+#' * `{y <- quo(x); arg_env_(y)}`
+#' * `arg_env_(quote(x), environment())`
+#' * `arg_env_(quo(x))`
+#' * `env(arg_(quo(x)))`.
 #' @rdname shortcut
 #' @param sym For plain `arg_env`, etc, a bare name, which is
 #'   quoted. For the underscore versions `arg_env_`, something that
@@ -76,21 +85,40 @@ dots_exprs <- function(...) {
 #'   optional names.
 is_forced <- function(...) {
   d <- dots(...)
-  is_forced_(exprs(d), envs(d))
+  is_forced_.default(exprs(d), envs(d))
 }
 
 #' @rdname shortcut
 #' @export
 #' @useDynLib nseval _is_forced
 is_forced_ <- function(syms, envs) {
+  UseMethod("is_forced_")
+}
+
+#' @exportS3Method is_forced_ default
+is_forced_.default <- function(syms, envs) {
   if (is.null(names(syms)))
     names(syms) <- as.character(syms)
   mapply(
     syms,
-    if(is.list(envs)) envs else list(envs),
+    if (is.list(envs)) envs else list(envs),
     FUN=function(sym, env) {
-      .Call("_is_forced", env, as.name(sym), TRUE)
+      .Call("_is_forced", env, as.name(sym), FALSE)
     })
+}
+
+#' @exportS3Method is_forced_ quotation
+is_forced_.quotation <- function(syms, ...) {
+  .Call("_is_forced", env(syms), expr(syms), FALSE)
+}
+
+#' @exportS3Method is_forced_ dots
+is_forced_.dots <- function(syms, ...) {
+  mapply(exprs(syms),
+         envs(syms),
+         FUN=function(sym, env) {
+           .Call("_is_forced", env, as.name(sym), FALSE)
+         })
 }
 
 #' @description `is_literal(x)` returns TRUE if an argument `x` could be a
@@ -101,7 +129,7 @@ is_forced_ <- function(syms, envs) {
 #' @export
 is_literal <- function(...) {
   d <- dots(...)
-  is_literal_(exprs(d), envs(d))
+  is_literal_.default(exprs(d), envs(d))
 }
 
 #' @rdname shortcut
@@ -109,26 +137,54 @@ is_literal <- function(...) {
 #' @param envs An environment or list of environments.
 #' @useDynLib nseval _is_literal
 is_literal_ <- function(syms, envs) {
+  UseMethod("is_literal_")
+}
+
+#' @exportS3Method is_literal_ default
+is_literal_.default <- function(syms, envs) {
   if (is.null(names(syms)))
     names(syms) <- as.character(syms)
   mapply(
     syms,
     if (is.list(envs)) envs else list(envs),
     FUN=function(sym, env) {
-      .Call("_is_literal", env, as.name(sym), TRUE)
+      .Call("_is_literal", env, as.name(sym), FALSE)
     })
 }
 
+#' @exportS3Method is_literal_ quotation
+is_literal_.quotation <- function(syms, ...) {
+  .Call("_is_literal", env(syms), expr(syms), FALSE)
+}
+
+#' @exportS3Method is_literal_ dots
+is_literal_.dots <- function(syms, ...) {
+  mapply(exprs(syms),
+         envs(syms),
+         FUN=function(sym, env) {
+           .Call("_is_literal", env, as.name(sym), FALSE)
+         })
+}
 
 #' @rdname shortcut
 #' @export
-#' @description `is_missing(...)` checks whether an argument is missing,
-#'   without forcing. It is similar to [missing] but can take
+#' @description `is_missing(...)` checks whether an argument is
+#'   missing, without forcing. It is similar to [missing] but can take
 #'   multiple arguments, and can be called in more situations, such as
 #'   from a nested inner function.
+#' @details When used with quotation objects, the `is_*_` functions
+#'   with trailing underscore work at one level of indirection
+#'   compared to quotation methods. For example, `missing_(x)` tests
+#'   whether `expr(x)` is `[missing_value()]`, whereas `is_missing_(x)`
+#'   assumes `expr(x)` is a _name_ and checks if that name refers to a
+#'   variable that is missing. The following are equivalent:
+#'
+#'   * `is_missing(x)`
+#'   * `is_missing_(quo(x))`
+#'   * `missing_(arg(x))`
 is_missing <- function(...) {
   d <- dots(...)
-  is_missing_(exprs(d), envs(d))
+  is_missing_.default(exprs(d), envs(d))
 }
 
 #' @rdname shortcut
@@ -137,15 +193,21 @@ is_missing <- function(...) {
 #'   missingness.
 #' @export
 #' @useDynLib nseval _is_missing
-#' @description `is_missing_(syms, envs)` is a normally evouating version
+#' @description `is_missing_(syms, envs)` is a normally evaluating version
 #'   of is_missing.  `syms` should be a symbol, character vector or
 #'   list of such. `envs` should be an environment, or list of
 #'   environments. Vector recycling rules apply, so you can call with
 #'   a vector of names and one env, or vice versa.
+#' @details When used with a `quotation` or `dots`, is_missing(q)
+#' looks for the variable(s) specified by expr(q) in environment env(q)]`.
 is_missing_ <- function(syms, envs, unwrap=TRUE) {
+  UseMethod("is_missing_")
+}
+
+#' @exportS3Method is_missing_ default
+is_missing_.default <- function(syms, envs, unwrap=TRUE) {
   if (is.null(names(syms)))
     names(syms) <- as.character(syms)
-  if (!is.list(envs)) envs <- list(envs)
   mapply(
     syms,
     if (is.list(envs)) envs else list(envs),
@@ -154,30 +216,65 @@ is_missing_ <- function(syms, envs, unwrap=TRUE) {
     })
 }
 
+#' @exportS3Method is_missing_ quotation
+#' @rdname shortcut
+is_missing_.quotation <- function(syms, ..., unwrap=TRUE) {
+  .Call("_is_missing", env(syms), expr(syms), FALSE, unwrap)
+}
+
+#' @exportS3Method is_missing_ dots
+is_missing_.dots <- function(syms, ..., unwrap=TRUE) {
+  mapply(exprs(syms),
+         envs(syms),
+         FUN=function(sym, env) {
+           .Call("_is_missing", env, as.name(sym), FALSE, unwrap)
+         })
+}
+
+
 #' @description `is_promise` returns TRUE if the given variable is bound to
 #'   a promise. Not all arguments are bound to promises; byte-compiled
 #'   code often omits creating a promise for literal or missing arguments.
 #' @rdname shortcut
 #' @export
-#' @param ... Bare variable names (for `is_*_`) or expressions (for
+#' @param ... Bare variable names (for `is_*`) or expressions (for
 #'   `dots_*`). Not forced.
 is_promise <- function(...) {
-  is_promise_(dots_exprs(...), dots_envs(...))
+  is_promise_.default(dots_exprs(...), dots_envs(...))
 }
 
 #' @rdname shortcut
 #' @export
 #' @useDynLib nseval _is_promise
-is_promise_ <- function(syms, envs)
-{
+is_promise_ <- function(syms, envs) {
+  UseMethod("is_promise_")
+}
+
+#' @exportS3Method is_promise_ default
+is_promise_.default <- function(syms, envs) {
   if (is.null(names(syms)))
     names(syms) <- as.character(syms)
   mapply(
     syms,
     if (is.list(envs)) envs else list(envs),
     FUN=function(sym, env) {
-      .Call("_is_promise", env, as.name(sym), TRUE)
+      .Call("_is_promise", env, as.name(sym), FALSE)
     })
+}
+
+#' @exportS3Method is_promise_ quotation
+#' @rdname shortcut
+is_promise_.quotation <- function(syms, ...) {
+  .Call("_is_promise", env(syms), expr(syms), FALSE)
+}
+
+#' @exportS3Method is_promise_ dots
+is_promise_.dots <- function(syms, ...) {
+  mapply(exprs(syms),
+         envs(syms),
+         FUN=function(sym, env) {
+           .Call("_is_promise", env, as.name(sym), FALSE)
+         })
 }
 
 #' @description
@@ -188,21 +285,50 @@ is_promise_ <- function(syms, envs)
 #' @rdname shortcut
 #' @export
 is_default <- function(...) {
-  is_default_(dots_exprs(...), dots_envs(...))
+  is_default_.default(dots_exprs(...), dots_envs(...))
 }
 
 #' @rdname shortcut
 #' @export
-is_default_ <- function(syms, envs = arg_env(syms)) {
-  if (is.null(names(syms)))
-    names(syms) <- as.character(syms)
-  mapply(
-    syms,
-    if (is.list(envs)) envs else list(envs),
-    FUN=function(sym, env) {
-      (    identical(arg_env_(sym, env),
-                     env)
-        && identical(arg_expr_(sym, env),
-                     formals(get_function(env))[[sym]]))
-    })
+is_default_ <- function(syms, envs) {
+  UseMethod("is_default_")
+}
+
+is_default__ <- function(sym, env) {
+  ( identical(arg_env_(sym, env),
+                 env)
+    && identical(arg_expr_(sym, env),
+                 formals(get_function(locate_(sym,env)))[[sym]]))
+}
+
+is_default__ <- function(sym, env) {
+  env <- locate_(sym, env)
+  (    identical(arg_env_(sym, env),
+                 env)
+    && identical(arg_expr_(sym, env),
+                 formals(get_function(env))[[sym]]))
+}
+
+
+#' @exportS3Method is_default_ default
+is_default_.default <- function(syms, envs) {
+    if (is.null(names(syms)))
+      names(syms) <- as.character(syms)
+    mapply(
+      syms,
+      if (is.list(envs)) envs else list(envs),
+      FUN=is_default__)
+}
+
+#' @exportS3Method is_default_ quotation
+#' @rdname shortcut
+is_default_.quotation <- function(syms, ...) {
+  is_default__(env(syms), expr(syms))
+}
+
+#' @exportS3Method is_default_ dots
+is_default_.dots <- function(syms, ...) {
+  mapply(exprs(syms),
+         envs(syms),
+         FUN=is_default__)
 }